combine parser and queries
