+++ queries.py
@@ -4,6 +4,19 @@ import logging
 import pandas
 
 
+def parse_csv_into_dataframe(filepath):
+    """ parses csv file into an pandas dataframe """
+    dataframe = pandas.read_csv(filepath)
+    logging.debug("generated pandas dataframe:")
+    logging.debug(dataframe)
+    return dataframe
+
+
+def _determine_latest(dataframe, short_email, date, field):
+    """ returns true if given field is the latest response """
+    # return: true/false
+
+
 def query_list(dataframe):
     """ returns list of emails of the advisees that have responded """
     # return: [email, email, ...]
@@ -28,8 +41,3 @@ def query_search(dataframe, keyword):
 def query_search_field(dataframe, field, keyword):
     """ returns dataframe rows where the field contains the keyword """
     # return: (keyword, field, [((timestamp, latest), email, response)])
-
-
-def determine_latest(dataframe, short_email, date, field):
-    """ returns true if given field is the latest response """
-    # return: true/false
